+++ cucumber-testng/src/test/java/io/cucumber/testng/UndefinedStepExceptionTest.java
@@ -2,48 +2,45 @@ package io.cucumber.testng;
 
 import io.cucumber.core.runtime.TestCaseResultObserver.Suggestion;
 import io.cucumber.plugin.event.Location;
-import org.testng.annotations.Test;
+import org.junit.jupiter.api.Test;
 
 import java.net.URI;
 
 import static java.util.Arrays.asList;
 import static java.util.Collections.emptyList;
 import static java.util.Collections.singletonList;
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.equalTo;
-import static org.hamcrest.Matchers.is;
-import static org.hamcrest.collection.IsArrayWithSize.arrayWithSize;
+import static org.assertj.core.api.Assertions.assertThat;
 
-public class UndefinedStepExceptionTest {
+class UndefinedStepExceptionTest {
 
     private final URI uri = URI.create("classpath:example.feature");
     private final Location stepLocation = new Location(12, 4);
 
     @Test
-    public void should_generate_a_message_for_no_suggestions() {
+    void should_generate_a_message_for_no_suggestions() {
         UndefinedStepException exception = new UndefinedStepException(emptyList());
-        assertThat(exception.getMessage(), is("This step is undefined"));
+        assertThat(exception).hasMessage("This step is undefined");
     }
 
     @Test
     void should_generate_an_empty_stacktrace_for_no_suggestions() {
         UndefinedStepException exception = new UndefinedStepException(emptyList());
-        assertThat(exception.getStackTrace(), arrayWithSize(0));
+        assertThat(exception.getStackTrace()).isEmpty();
     }
 
     @Test
-    public void should_generate_a_message_for_one_suggestions() {
+    void should_generate_a_message_for_one_suggestions() {
         UndefinedStepException exception = new UndefinedStepException(
             singletonList(
                 new Suggestion("some step", singletonList("some snippet"), uri, stepLocation))
 
         );
-        assertThat(exception.getMessage(), is("""
+        assertThat(exception).hasMessage("""
                 The step 'some step' is undefined.
                 You can implement this step using the snippet(s) below:
 
                 some snippet
-                """));
+                """);
     }
 
     @Test
@@ -53,29 +50,30 @@ public class UndefinedStepExceptionTest {
                 new Suggestion("some step", singletonList("some snippet"), uri, stepLocation))
 
         );
-        assertThat(exception.getStackTrace(), arrayWithSize(1));
-        assertThat(exception.getStackTrace()[0].toString(), equalTo("✽.some step(classpath:example.feature:12)"));
+        assertThat(exception.getStackTrace())
+                .extracting(StackTraceElement::toString)
+                .containsExactly("✽.some step(classpath:example.feature:12)");
     }
 
     @Test
-    public void should_generate_a_message_for_one_suggestions_with_multiple_snippets() {
+    void should_generate_a_message_for_one_suggestions_with_multiple_snippets() {
         UndefinedStepException exception = new UndefinedStepException(
             singletonList(
                 new Suggestion("some step", asList("some snippet", "some other snippet"), uri,
                     stepLocation))
 
         );
-        assertThat(exception.getMessage(), is("""
+        assertThat(exception).hasMessage("""
                 The step 'some step' is undefined.
                 You can implement this step using the snippet(s) below:
 
                 some snippet
                 some other snippet
-                """));
+                """);
     }
 
     @Test
-    public void should_generate_a_message_for_two_suggestions() {
+    void should_generate_a_message_for_two_suggestions() {
         UndefinedStepException exception = new UndefinedStepException(
             asList(
                 new Suggestion("some step", singletonList("some snippet"), uri, stepLocation),
@@ -83,17 +81,17 @@ public class UndefinedStepExceptionTest {
                     stepLocation))
 
         );
-        assertThat(exception.getMessage(), is("""
+        assertThat(exception).hasMessage("""
                 The step 'some step' and 1 other step(s) are undefined.
                 You can implement these steps using the snippet(s) below:
 
                 some snippet
                 some other snippet
-                """));
+                """);
     }
 
     @Test
-    public void should_generate_a_message_without_duplicate_suggestions() {
+    void should_generate_a_message_without_duplicate_suggestions() {
         UndefinedStepException exception = new UndefinedStepException(
             asList(
                 new Suggestion("some step", asList("some snippet", "some snippet"), uri,
@@ -102,17 +100,17 @@ public class UndefinedStepExceptionTest {
                     stepLocation))
 
         );
-        assertThat(exception.getMessage(), is("""
+        assertThat(exception).hasMessage("""
                 The step 'some step' and 1 other step(s) are undefined.
                 You can implement these steps using the snippet(s) below:
 
                 some snippet
                 some other snippet
-                """));
+                """);
     }
 
     @Test
-    public void should_generate_a_message_for_three_suggestions() {
+    void should_generate_a_message_for_three_suggestions() {
         UndefinedStepException exception = new UndefinedStepException(
             asList(
                 new Suggestion("some step", singletonList("some snippet"), uri, stepLocation),
@@ -122,14 +120,14 @@ public class UndefinedStepExceptionTest {
                     stepLocation))
 
         );
-        assertThat(exception.getMessage(), is("""
+        assertThat(exception).hasMessage("""
                 The step 'some step' and 2 other step(s) are undefined.
                 You can implement these steps using the snippet(s) below:
 
                 some snippet
                 some other snippet
                 yet another snippet
-                """));
+                """);
     }
 
 }